+++ core/src/main/java/feign/ReflectiveFeign.java
@@ -1,5 +1,5 @@
 /**
- * Copyright 2012-2018 The Feign Authors
+ * Copyright 2012-2019 The Feign Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
  * in compliance with the License. You may obtain a copy of the License at
@@ -162,8 +162,11 @@ public class ReflectiveFeign extends Feign {
         } else {
           buildTemplate = new BuildTemplateByResolvingArgs(md, queryMapEncoder);
         }
+        Options requestOptions = md.options()
+            .map(o -> new Options(o.getConnectTimeoutMillis(), o.getReadTimeoutMillis()))
+            .orElse(options);
         result.put(md.configKey(),
-            factory.create(key, md, buildTemplate, options, decoder, errorDecoder));
+            factory.create(key, md, buildTemplate, requestOptions, decoder, errorDecoder));
       }
       return result;
     }