+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Unwrap RetryableException and throw cause (#737)
 * Supports PATCH without a body paramter (#824)
 * Feign-Ribbon integration now depends on Ribbon 2.3.0, updated from Ribbon 2.1.1 (#826)
+* Add support for per request timeout options with @ApiTimeout (#562)
 
 ### Version 10.0
 * Feign baseline is now JDK 8